Allow use_resource to be change per-row
`use_resource` allows you to use a non-standard
resource for an association. However, if you have
a polymorphic association you might want to use
a different resource for each row (eg for different
fields).

I've made this possible by allowing a block or lambda
to be passed to `use_resource`, which is evaluated
with the model for that row.